+++ tests/testthat/test-recipe-read.R
@@ -163,7 +163,6 @@ test_that("resource case matters", {
                "Script file does not exist: 'script.R'")
 })
 
-
 test_that("dependencies must be scalar", {
   path <- prepare_orderly_example("depends")
   id <- orderly_run("example", config = path, echo = FALSE)
@@ -189,4 +188,13 @@ test_that("dependencies must exist", {
 
   expect_error(orderly_run("depend", config = path, echo = FALSE),
                "Did not find file unknown.file at")
+})
+
+
+test_that("data field is optional", {
+  path <- prepare_orderly_example("nodata")
+  report_path <- file.path(path, "src", "example")
+
+  ## expect no error
+  expect_error(orderly_run("example", config = path, echo = FALSE), NA)
 })